Restructure Components into a Library
