+++ src/app/components/common/index.ts
@@ -1,0 +1,1 @@
+export * from './private-route/private-route.component';
