+++ src/app/components/example/example/example.component.tsx
