+++ src/app/components/index.ts
@@ -1,0 +1,4 @@
+export * from './common';
+export * from './example';
+export * from './pdf-display';
+export * from './policies';
