+++ src/app/components/pdf-display/document/document.component.tsx
