+++ src/app/components/pdf-display/index.ts
@@ -1,0 +1,3 @@
+export * from './document/document.component';
+export * from './page/page.component';
+export * from './pdf-display/pdf-display.component';
