+++ src/app/components/policies/index.ts
@@ -1,0 +1,2 @@
+export * from './privacy/privacy.component';
+export * from './tos/tos.component';
