+++ src/app/components/policies/privacy/privacy.component.tsx
