+++ src/app/components/policies/tos/tos.component.tsx
